llm evaluation

+++ src/prompt_llm/user_sim.py
@@ -1,10 +1,12 @@
 import click
 import json
-import glob
 import random
 import os
 import shutil
 import sys
+
+from tqdm import tqdm
+
 
 class TurnMaker:
     def __init__(self, **kwargs):
@@ -12,7 +14,7 @@ class TurnMaker:
         self.commander_name = kwargs["commander_name"]
         self.driver_name = kwargs['driver_name']
         self.n = kwargs["n"]
-        self.task_length = kwargs["length"]
+        self.task_length = eval(str(kwargs["length"]))
         self.numerate = kwargs["numerate"]
         self.save_path = kwargs["save_path"]
         self.give_task = not kwargs["no_task"]
@@ -26,6 +28,10 @@ class TurnMaker:
         self.cheat = kwargs["cheat"]
         self.ex_max_obs = kwargs["example_max_percent_observe"]
         self.entire_file = kwargs["entire_file"]
+        self.no_obs = kwargs["no_obs"]
+        self.split_dataset = kwargs["split_dataset"]
+        self.train = kwargs["train"]
+        self.entire_dataset = kwargs["entire_dataset"]
 
         if self.nl_ify:
             self.das = {
@@ -72,13 +78,36 @@ class TurnMaker:
 
         self.kwargs = kwargs
 
-        self.tasks: list[dict] = [task for datafile in glob.glob(os.path.join(self.data_path, "*_turn.json"))
-                                  for task in json.load(open(datafile, 'r'))]
+        # Chosen since it has the most examples
+        with open(os.path.join(self.data_path, "train_turn.json"), 'r') as f:
+            train_source: list[dict] = [task for task in json.load(f)]
+
+        # Chosen since it has the least number of examples
+        with open(os.path.join(self.data_path, "valid_seen_turn.json"), 'r') as f:
+            valid_seen_source: list[dict] = [task for task in json.load(f)]
+
+        with open(os.path.join(self.data_path, "valid_unseen_turn.json"), 'r') as f:
+            valid_unseen_source: list[dict] = [task for task in json.load(f)]
+
+        self.example_source: list[dict] = train_source + valid_unseen_source
+
+        if self.train:
+            self.task_source: list[dict] = train_source + valid_unseen_source
+        else:
+            self.task_source: list[dict] = valid_seen_source
+
+        self.tasks: list[dict] = self.example_source + self.task_source
 
     def agent_name(self, original_name) -> str:
-        return self.commander_name if original_name.lower() == "commander" else self.driver_name
+        match original_name.lower():
+            case "commander":
+                return self.commander_name
+            case "driver":
+                return self.driver_name
+            case _:
+                raise ValueError(f"Invalid agent name: {original_name}")
 
-    def process_turn(self, turn: dict, is_last) -> str:
+    def process_turn(self, turn: dict, is_last=False) -> str:
         driver_action = turn['DRIVER']['action'] if turn['DRIVER']['action'] != 'dialogue' else turn["DRIVER"]['utterance']
         commander_action = turn['COMMANDER']['action'] if turn['COMMANDER']['action'] != 'dialogue' else turn["COMMANDER"]['utterance']
         if turn['DRIVER']['action'] == 'dialogue':
@@ -90,12 +119,25 @@ class TurnMaker:
         else:
             commander_das = ""
         if self.html_format:
-            return (f"<TURN> <{self.commander_name}> {commander_action} {f'<<{commander_das}>> ' if self.include_das and commander_das else ''}</{self.commander_name}>\n"
-                    f"       <{self.driver_name}> {driver_action} {f'<<{driver_das}>> ' if self.include_das and driver_das else ''}</{self.driver_name}> </TURN>")
+            to_ret = "<TURN> "
+            if not (("observe" in commander_action.lower()) and self.no_obs):  # Either none or both
+                to_ret += f"<{self.commander_name}> {commander_action} {f'<<{commander_das}>> ' if self.include_das and commander_das else ''}</{self.commander_name}>"
+                if is_last:
+                    return to_ret
+            if not (("observe" in driver_action.lower()) and self.no_obs):
+                to_ret += f"<{self.driver_name}> {driver_action} {f'<<{driver_das}>> ' if self.include_das and driver_das else ''}</{self.driver_name}>"
+            to_ret += "</TURN>"
         else:
-            return f"{self.commander_name}: {commander_action}{f' <<{commander_das}>>'if (self.include_das or (self.predict_das and is_last)) and commander_das else ''}\n{self.driver_name}: {driver_action}{f' <<{driver_das}>>' if self.include_das and driver_das else ''}"
+            to_ret = ""
+            if not (("observe" in commander_action.lower()) and self.no_obs):  # Either none or both
+                to_ret += f"{self.commander_name}: {commander_action}{f' <<{commander_das}>>' if self.include_das and commander_das else ''}" + "\n" * (not self.no_obs)
+                if is_last:
+                    return to_ret
+            if not (("observe" in driver_action.lower()) and self.no_obs):
+                to_ret += f"{self.driver_name}: {driver_action}{f' <<{driver_das}>>' if self.include_das and driver_das else ''}"
+        return to_ret
 
-    def get_rand_turns(self, task: dict, get_last=True, length=None) -> list[str]:
+    def get_rand_turns(self, task: dict, get_last=True, length=None) -> list[dict]:
         if length is not None:
             length = length
         elif len(task['turns']) < self.min_length:
@@ -104,7 +146,7 @@ class TurnMaker:
             length = self.task_length if self.task_length is not None else random.randint(self.min_length, min(len(task['turns']), self.max_length))
         start = 0
         end = start + length
-        return [self.process_turn(turn, get_last and (i == length - 1)) for i, turn in enumerate(task['turns'][start:end])]
+        return task['turns'][start:end]
 
     def display_examples(self) -> str:
         examples = []
@@ -131,21 +173,20 @@ class TurnMaker:
         else:
             prompt = f"Goal: {task['goal']}\n"
         turns = self.get_rand_turns(task, length=length)
-        prompt += "\n".join(turns[:-1]) + '\n'
+        prompt += "\n".join(map(self.process_turn, turns[:-1])) + '\n'
         prompt += f"{self.commander_name} response:"
-        ans = turns[-1].splitlines()[0]
+        ans = self.process_turn(turns[-1], is_last=True)
         if self.html_format:
             ans = ans[ans.find(f"<{self.commander_name}>")+2+len(self.commander_name):ans.find("</")]
         else:
-            ans = ans.split(":")[1].strip()
+            ans = ans.split(":", maxsplit=1)[1].strip()
         if not self.predict_das:
             act = "OBSERVE" if "<observe>" in ans else "SPEAK"
         else:
             try:
                 act = "OBSERVE" if "observe" in ans.lower() else ans[ans.index("<<") + 2:ans.index(">>")].split(",")[0].strip()
             except ValueError:
-                print(ans)
-                sys.exit()
+                raise ValueError(f"Error in parsing dialogue act: {task['file_id']} {turns[-1]}")
         utt = ans.split(":")[-1][:ans.rfind("<<")].strip() if self.gen_response and "<observe>" not in ans else ""
         return prompt, f"{act}\n{utt}".strip()
 
@@ -177,21 +218,25 @@ class TurnMaker:
         answer = ""
         if self.give_task and not override:
             prompt += "Give your answer for the following example:\n"
-            task = random.choice(self.tasks)
+            task = random.choice(self.example_source if self.split_dataset else self.tasks)
             p, a = self.display_example(task)
             prompt += p + '\n'
             answer = a
 
         return prompt, answer
 
-    def generate_file(self, save_answer, save_path):
-        task = random.choice(self.tasks)
+    def generate_file(self, save_answer, save_path, task: dict = None, dont_edit_folders=False, dont_print=False):
+        if task is None:
+            task = random.choice(self.tasks)
         save_root, save_ext = save_path.rsplit(".", 1)
         episodes = [self.display_example(task, i) for i in range(1, len(task['turns'])+1)]
         prompts = [self.make_prompt(override=True) for _ in range(len(episodes))]
-        if os.path.exists(save_path[:save_path.rfind('/')]):
-            shutil.rmtree(save_path[:save_path.rfind('/')])
-        os.mkdir(save_path[:save_path.rfind('/')])
+
+        if not dont_edit_folders:
+            if os.path.exists(save_path[:save_path.rfind('/')]):
+                shutil.rmtree(save_path[:save_path.rfind('/')])
+            os.mkdir(save_path[:save_path.rfind('/')])
+
         for i, ((prompt, _), (task, answer)) in enumerate(zip(prompts, episodes)):
             with open(f"{save_root}_{i}.{save_ext}", "w+") as f:
                 f.write(prompt)
@@ -199,7 +244,24 @@ class TurnMaker:
                 if save_answer:
                     with open(f"{save_root}_{i}_answer.{save_ext}", "w+") as a:
                         a.write(answer)
-        print(f"Saved to {save_path[:save_path.rfind('/')]}")
+        if not dont_print:
+            print(f"Saved to {save_path[:save_path.rfind('/')]}")
+
+    def generate_entire_dataset(self, save_path):
+        save_root = save_path
+        save_file = "turn.txt"
+
+        if os.path.exists(save_root):
+            shutil.rmtree(save_root)
+        os.mkdir(save_root)
+
+        print("Generating entire dataset...")
+        for target in tqdm(self.task_source):
+            folder = os.path.join(save_root, target["file_id"])
+            os.mkdir(folder)
+            path = os.path.join(folder, save_file)
+            self.generate_file(True, path, task=target, dont_edit_folders=True, dont_print=True)
+
 
     def generate_prompt(self, save_answer=False, save_path=None, num_to_gen=1):
         if save_path is None:
@@ -251,11 +313,11 @@ class TurnMaker:
 @click.option("--data_path", "-d", default="teach-dataset-parsed/", help="Path to the data")
 @click.option("--commander_name", default="COMMANDER", help="Name of the commander")
 @click.option("--driver_name", default="DRIVER", help="Name of the driver")
-@click.option("--n", "-n", "--num", default=2, help="Number of examples to generate")
+@click.option("--n", "-n", "--num", default=5, help="Number of examples to generate")
 @click.option("--length", "-l", default=None, help="Length of the examples")
 @click.option("--numerate", is_flag=True, help="Whether to number the examples")
 @click.option("--save_path", "-s", default="src/prompt_llm/generated_turn.txt", help="Path to save the generated prompt")
-@click.option("--max_length", default=25, help="Maximum length of the examples")
+@click.option("--max_length", default=10, help="Maximum length of the examples")
 @click.option("--min_length", default=1, help="Minimum length of the examples")
 @click.option("--no_task", "-t", is_flag=True, help="Whether to give a task to the LLM to classify at the end")
 @click.option("--include_das", "-i", is_flag=True, help="Whether to include dialogue acts in the examples")
@@ -267,18 +329,21 @@ class TurnMaker:
 @click.option("--example_max_percent_observe", default=35, help="Maximum percentage of given examples to be observation (to balance the data)")
 @click.option("--max_percent_observe", default=35, help="Maximum percentage of tasks to be observation (to balance the data)")
 @click.option("--gen_response", "-r", is_flag=True, help="Whether to generate a response to the prompt in addition to the dialogue act")
-# TODO add flag for observes to be included on
+# TODO setup running with images
 # TODO consider visual features (don't have them yet; nl-ified)
 # TODO add flag on standard (random) vs cheating (don't make the only example have the same answer as the examples)
 @click.option("--cheat", "-c", is_flag=True, help="Whether to cheat and make the only example have the same answer as the examples")
-# TODO consider adding two examples with the same goal (maybe even from the same scenario)
-# TODO measure using accuracy (split up) and f measure ->
-# TODO get accuracy for entire game file sequentially
+@click.option("--no_obs", is_flag=True, help="Whether to omit observes in the examples provided, but not the task")
 @click.option("--entire_file", "-e", is_flag=True, help="Whether to generate prompts for an entire game file")
+@click.option("--split_dataset", is_flag=True, help="Whether to split the dataset")
+@click.option("--train", is_flag=True, help="Whether to generate prompts for the training set, not the entire dataset")
+@click.option("--entire_dataset", is_flag=True, help="Whether to generate prompts for the entire dataset. Save path should be a folder for this one.")
 def main(**kwargs):
     tm = TurnMaker(**kwargs)
-    tm.generate_prompt(save_answer=kwargs["save_answer"], num_to_gen=kwargs["num_prompts"])
-    # todo try 20-30 examples
+    if kwargs["entire_dataset"]:
+        tm.generate_entire_dataset(kwargs["save_path"])
+    else:
+        tm.generate_prompt(save_answer=kwargs["save_answer"], num_to_gen=kwargs["num_prompts"])
 
 if __name__ == "__main__":
     main()
